+++ CHANGELOG.md
@@ -11,14 +11,22 @@ Features:
   location - #122
 - Closing a tab or a window that still has a running process will now prompt you
   for confirmation - fixes #201
-- Black Box now uses the default Adwaita tab style - #112
+- Black Box now uses the default Adwaita tab style. As consequence, some header
+  bar options, including "Show Borders" have been removed from the preferences
+  window - #112, #253
+- Added the option to disable terminal bell - #106
+
+Improvements:
+
+- Some configuration options have been grouped together in the preferences
+  window - #254
+- Application title is now bold when there's a single tab open - #235
 
 Bug fixes:
 
+- The window title is now centered when there's only one tab - #199
 - Improved keybinding validation, allowing more valid key combinations to be
   used - #245
-
-> Nothing
 
 ## 0.13.2 - 2023-01-19